refactor: split expression conversion
